+++ include/info.h
@@ -1,4 +1,5 @@
 /*
+ * Copyright 2013 Esrille Inc.
  * Copyright 2008-2010 Google Inc.
  * Copyright 2007 Nintendo Co., Ltd.
  *
@@ -77,7 +78,7 @@ public:
 
     virtual void at(const OpDcl* node)
     {
-        for (int i = 0; i < node->getMethodCount(); ++i)
+        for (size_t i = 0; i < node->getMethodCount(); ++i)
         {
             if (i != 0)
             {
