+++ include/messenger.h
@@ -559,7 +559,7 @@ public:
         {
             included = "No" + included;
         }
-        for (int i = 0; i < included.size(); ++i)
+        for (size_t i = 0; i < included.size(); ++i)
         {
             char c = included[i];
             included[i] = isalnum(c) ? (capitalize ? toupper(c) : c) : '_';
