+++ src/cplusplus.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 Esrille Inc.
+ * Copyright 2010-2013 Esrille Inc.
  * Copyright 2008-2010 Google Inc.
  * Copyright 2007 Nintendo Co., Ltd.
  *
@@ -633,17 +633,17 @@ public:
             pos += 2;
         }
 
-        int i;
+        size_t i;
         for (i = 0;
              i < prevName.size() && i < packageName.size() && prevName[i] == packageName[i];
              ++i)
         {
         }
-        for (int j = i; j < prevName.size(); ++j)
+        for (size_t j = i; j < prevName.size(); ++j)
         {
             writeln("}");
         }
-        for (int j = i; j < packageName.size(); ++j)
+        for (size_t j = i; j < packageName.size(); ++j)
         {
             writeln("namespace %s {", CPlusPlus::getEscapedName(packageName[j]).c_str());
         }
@@ -656,7 +656,7 @@ public:
         if (0 < packageName.size())
         {
             write("\n");
-            for (int i = 0; i < packageName.size(); ++i)
+            for (size_t i = 0; i < packageName.size(); ++i)
             {
                 writeln("}");
             }
